changed from m floats to ints

+++ src/main/scala/ayai/components/Velocity.scala
@@ -2,5 +2,5 @@ package ayai.components
 
 import com.artemis.Component
 
-class Velocity (var x : Float, var y : Float) extends Component {
+class Velocity (var x : Int, var y : Int) extends Component {
 }
